+++ lib/BeGateway/PaymentMethod/CreditCard.php
@@ -3,4 +3,3 @@ namespace BeGateway\PaymentMethod;
 
 class CreditCard extends Base {
 }
-?>